Move HttpClientInterface types out of Types.d.ts
